+++ NEWS.md
@@ -38,7 +38,21 @@
   .data[["name"]])` and `group_by(df, name)` produce the same column
   name.
 
-* `call_modify()` supports removing unexisting arguments (#393).
+* The `.standardise` argument of `call_modify()` is deprecated. Please
+  use `call_standardise()` beforehand.
+
+* `call_modify()` has better support for `...` and now treats it like
+  a named argument. `call_modify(call, ... = )` adds `...` to the call
+  and `call_modify(call, ... = NULL)` removes it.
+
+* `call_modify()` now preserves empty arguments. It is no longer
+  necessary to use `missing_arg()` to add a missing argument to a
+  call. This is possible thanks to the new `.preserve_empty` option of
+  `dots_list()`.
+
+* `call_modify()` now supports removing unexisting arguments (#393)
+  and passing multiple arguments with the same name (#398). The new
+  `.homonyms` argument controls how to treat these arguments.
 
 * `call_standardise()` now handles primitive functions like `~`
   properly (#473).